Separate index and bin concerns
